requirements.txt
numpy
matplotlib
click
pyzmq
uvicorn
python-multipart
torch
faiss-cpu
sentence-transformers
gradio
